+++ plot/viewer/src/canvases/SketchCanvas.tsx
@@ -379,7 +379,6 @@ function SketchCanvasInner({
     handleNodesDelete,
     handleEdgesDelete,
     addNodeAt,
-    selectedNodeIds,
   });
 
   useKeyboardShortcuts({